+++ compass/app/auth/forgot_password/page.tsx
@@ -1,5 +1,5 @@
 "use client"
-import Button from '@/components/Button49';
+import Button from '@/components/Button';
 import Input from '@/components/Input'
 import InlineLink from '@/components/InlineLink';
 import Paper from '@/components/auth/Paper';
@@ -41,10 +41,10 @@ export default function Page()  {
                         title="Enter your email address"
                         placeholder="janedoe@gmail.com"
                         value={confirmEmail}
-                        iconKey={'EmailInputIcon'}
+                        icon={'EmailInputIcon'}
                         onChange={(e) => {
                             setconfirmEmail(e.target.value);
-                            setEmailError(''); // Reset the error when the user types
+                            // setEmailError(''); // Reset the error when the user types
                         }}/>
                         {emailError && (
                         <p className={`mt-2 ${emailError.includes('exists in the database') ? 'text-green-500' : 'text-red-500'}`}>
